+++ vkconfig_core/command_line.cpp
@@ -38,7 +38,8 @@ struct CommandHelpDesc {
 };
 
 static const CommandHelpDesc command_help_desc[] = {
-    {HELP_HELP, "help"}, {HELP_VERSION, "version"}, {HELP_LAYERS, "layers"}, {HELP_DOC, "doc"}, {HELP_RESET, "reset"},
+    {HELP_HELP, "help"},     {HELP_VERSION, "version"}, {HELP_LAYERS, "layers"},
+    {HELP_LOADER, "loader"}, {HELP_DOC, "doc"},         {HELP_RESET, "reset"},
 };
 
 static HelpType GetCommandHelpId(const char* token) {
@@ -67,6 +68,8 @@ static const ModeDesc mode_desc[] = {
     {COMMAND_VERSION, "version", HELP_VERSION},    // COMMAND_VERSION
     {COMMAND_LAYERS, "--layers", HELP_LAYERS},     // COMMAND_LAYERS
     {COMMAND_LAYERS, "layers", HELP_LAYERS},       // COMMAND_LAYERS
+    {COMMAND_LOADER, "--loader", HELP_LOADER},     // COMMAND_LOADER
+    {COMMAND_LOADER, "loader", HELP_LOADER},       // COMMAND_LOADER
     {COMMAND_DOC, "--doc", HELP_DOC},              // COMMAND_DOC
     {COMMAND_DOC, "doc", HELP_DOC},                // COMMAND_DOC
     {COMMAND_GUI, "--gui", HELP_GUI},              // COMMAND_GUI
@@ -124,12 +127,29 @@ struct CommandLayersDesc {
 static const CommandLayersDesc command_layers_desc[] = {
     {COMMAND_LAYERS_LIST, "--list", 2},
     {COMMAND_LAYERS_LIST, "-l", 2},
+    {COMMAND_LAYERS_PATH, "--path", 2},
+    {COMMAND_LAYERS_PATH, "-p", 2},
     {COMMAND_LAYERS_VERBOSE, "--list-verbose", 2},
     {COMMAND_LAYERS_VERBOSE, "-lv", 2},
     {COMMAND_LAYERS_OVERRIDE, "--override", 3},
     {COMMAND_LAYERS_OVERRIDE, "-o", 3},
     {COMMAND_LAYERS_SURRENDER, "--surrender", 2},
     {COMMAND_LAYERS_SURRENDER, "-s", 2},
+};
+
+struct CommandLoaderDesc {
+    CommandLoaderArg arguments;
+    const char* token;
+    int required_arguments;
+};
+
+static const CommandLoaderDesc command_loader_desc[] = {
+    {COMMAND_LOADER_LIST, "--list", 2},           {COMMAND_LOADER_LIST, "-l", 2},
+    {COMMAND_LOADER_OVERRIDE, "--override", 3},   {COMMAND_LOADER_OVERRIDE, "-o", 3},
+    {COMMAND_LOADER_SURRENDER, "--surrender", 2}, {COMMAND_LOADER_SURRENDER, "-s", 2},
+    {COMMAND_LOADER_IMPORT, "--import", 3},       {COMMAND_LOADER_IMPORT, "-i", 3},
+    {COMMAND_LOADER_EXPORT, "--export", 4},       {COMMAND_LOADER_EXPORT, "-e", 4},
+    {COMMAND_LOADER_DELETE, "--delete", 3},       {COMMAND_LOADER_DELETE, "-d", 3},
 };
 
 struct CommandDocDesc {
@@ -165,6 +185,27 @@ static const CommandLayersDesc& GetCommandLayers(CommandLayersArg layers_arg) {
     return command_layers_desc[0];
 }
 
+static CommandLoaderArg GetCommandLoaderId(const char* token) {
+    assert(token != nullptr);
+
+    for (std::size_t i = 0, n = std::size(command_loader_desc); i < n; ++i) {
+        if (std::strcmp(command_loader_desc[i].token, token) == 0) return command_loader_desc[i].arguments;
+    }
+
+    return COMMAND_LOADER_NONE;
+}
+
+static const CommandLoaderDesc& GetCommandLoader(CommandLoaderArg loader_arg) {
+    assert(loader_arg != COMMAND_LOADER_NONE);
+
+    for (std::size_t i = 0, n = std::size(command_loader_desc); i < n; ++i) {
+        if (command_loader_desc[i].arguments == loader_arg) return command_loader_desc[i];
+    }
+
+    assert(0);
+    return command_loader_desc[0];
+}
+
 static CommandDocArg GetCommandDocId(const char* token) {
     assert(token != nullptr);
 
@@ -179,6 +220,8 @@ CommandLine::CommandLine(int argc, char* argv[])
     : command(_command),
       command_reset_arg(_command_reset_arg),
       command_layers_arg(_command_layers_arg),
+      command_loader_arg(_command_loader_arg),
+      layers_configuration_name(_layers_configuration_name),
       layers_configuration_path(_layers_configuration_path),
       command_doc_arg(_command_doc_arg),
       doc_layer_name(_doc_layer_name),
@@ -187,6 +230,7 @@ CommandLine::CommandLine(int argc, char* argv[])
       error_args(_error_args),
       _command_reset_arg(COMMAND_RESET_NONE),
       _command_layers_arg(COMMAND_LAYERS_NONE),
+      _command_loader_arg(COMMAND_LOADER_NONE),
       _command_doc_arg(COMMAND_DOC_NONE),
       _error(ERROR_NONE),
       _help(HELP_DEFAULT) {
@@ -196,6 +240,72 @@ CommandLine::CommandLine(int argc, char* argv[])
     int arg_offset = 1;
 
     switch (_command = GetModeId(argv[arg_offset + 0])) {
+        case COMMAND_LOADER: {
+            if (argc <= arg_offset + 1) {
+                _error = ERROR_MISSING_COMMAND_ARGUMENT;
+                _error_args.push_back(argv[arg_offset + 0]);
+                break;
+            }
+
+            _command_loader_arg = GetCommandLoaderId(argv[arg_offset + 1]);
+            if (_command_loader_arg == COMMAND_LOADER_NONE) {
+                _error = ERROR_INVALID_COMMAND_ARGUMENT;
+                _error_args.push_back(argv[arg_offset + 0]);
+                _error_args.push_back(argv[arg_offset + 1]);
+                break;
+            }
+
+            const CommandLoaderDesc& desc = GetCommandLoader(_command_loader_arg);
+            if (argc < arg_offset + desc.required_arguments) {
+                _error = ERROR_MISSING_COMMAND_ARGUMENT;
+                _error_args.push_back(argv[arg_offset + 0]);
+                break;
+            } else if (argc > arg_offset + desc.required_arguments) {
+                _error = ERROR_TOO_MANY_COMMAND_ARGUMENTS;
+                _error_args.push_back(argv[arg_offset + 0]);
+                break;
+            }
+
+            switch (_command_loader_arg) {
+                default:
+                    break;
+                case COMMAND_LOADER_OVERRIDE: {
+                    const std::string layers_configuration = argv[arg_offset + 2];
+
+                    QFile file(Path(layers_configuration).AbsolutePath().c_str());
+                    const bool result = file.open(QFile::ReadOnly);
+                    if (!result) {
+                        //_error = ERROR_FILE_NOTFOUND;
+                        //_error_args.push_back(argv[arg_offset + 2]);
+                        _layers_configuration_name = layers_configuration;
+                    } else {
+                        _layers_configuration_path = layers_configuration;
+                    }
+                } break;
+                case COMMAND_LOADER_IMPORT: {
+                    _layers_configuration_path = argv[arg_offset + 2];
+
+                    QFile file(_layers_configuration_path.AbsolutePath().c_str());
+                    const bool result = file.open(QFile::ReadOnly);
+                    if (!result) {
+                        _error = ERROR_FILE_NOTFOUND;
+                        _error_args.push_back(argv[arg_offset + 2]);
+                    }
+                } break;
+                case COMMAND_LOADER_EXPORT: {
+                    _layers_configuration_name = argv[arg_offset + 2];
+                    _layers_configuration_path = argv[arg_offset + 3];
+                    Path ExportDir(_layers_configuration_path.AbsoluteDir());
+                    ExportDir.Create();
+                    if (!ExportDir.Exists()) {
+                        _error = ERROR_FILE_NOTFOUND;
+                    }
+                } break;
+                case COMMAND_LOADER_DELETE: {
+                    _layers_configuration_name = argv[arg_offset + 2];
+                } break;
+            }
+        } break;
         case COMMAND_LAYERS: {
             if (argc <= arg_offset + 1) {
                 _error = ERROR_MISSING_COMMAND_ARGUMENT;
@@ -224,7 +334,7 @@ CommandLine::CommandLine(int argc, char* argv[])
 
             if (_command_layers_arg == COMMAND_LAYERS_OVERRIDE) {
                 _layers_configuration_path = argv[arg_offset + 2];
-                QFile file(_layers_configuration_path.c_str());
+                QFile file(_layers_configuration_path.AbsolutePath().c_str());
                 const bool result = file.open(QFile::ReadOnly);
                 if (!result) {
                     _error = ERROR_FILE_NOTFOUND;
@@ -313,31 +423,31 @@ void CommandLine::log() const {
         } break;
         case ERROR_INVALID_COMMAND_USAGE: {
             assert(_error_args.size() == 1);
-            printf("Invalid '%s' command usage...\n\n", _error_args[0].c_str());
+            printf("vkconfig: Invalid '%s' command usage...\n\n", _error_args[0].c_str());
         } break;
         case ERROR_INVALID_COMMAND_ARGUMENT: {
             assert(_error_args.size() == 2);
-            printf("Invalid '%s' command argument: '%s'...\n\n", _error_args[0].c_str(), _error_args[1].c_str());
+            printf("vkconfig: Invalid '%s' command argument: '%s'...\n\n", _error_args[0].c_str(), _error_args[1].c_str());
         } break;
         case ERROR_MISSING_COMMAND_ARGUMENT: {
             assert(_error_args.size() == 1);
-            printf("Missing '%s' command argument...\n\n", _error_args[0].c_str());
+            printf("vkconfig: Missing '%s' command argument...\n\n", _error_args[0].c_str());
         } break;
         case ERROR_TOO_MANY_COMMAND_ARGUMENTS: {
             assert(_error_args.size() == 1);
-            printf("Too many '%s' command arguments...\n\n", _error_args[0].c_str());
+            printf("vkconfig: Too many '%s' command arguments...\n\n", _error_args[0].c_str());
         }
         case ERROR_FILE_NOTFOUND: {
             assert(_error_args.size() == 1);
-            printf("'%s' couldn't be found...\n\n", _error_args[0].c_str());
+            printf("vkconfig: '%s' couldn't be found...\n\n", _error_args[0].c_str());
         } break;
         case ERROR_UNKNOWN_ARGUMENT: {
             assert(_error_args.size() == 1);
-            printf("Unknown argument: '%s'...\n\n", _error_args[0].c_str());
+            printf("vkconfig: Unknown argument: '%s'...\n\n", _error_args[0].c_str());
         } break;
         default: {
             assert(0);
-            printf("Unknown error...\n\n");
+            printf("vkconfig: Unknown error...\n\n");
         } break;
     }
 }
@@ -350,17 +460,18 @@ void CommandLine::usage() const {
         }
         case HELP_DEFAULT: {
             printf("Usage\n");
-            printf("\tvkconfig [[help] | [version] | [layers <args>] | [reset] | [doc]]\n");
+            printf("\tvkconfig [[help] | [version] | [gui] | [layers <args>] | [loader <args>] | [reset] | [doc]]\n");
             printf("\n");
             printf("Command:\n");
             printf("\thelp                      = Display usage and documentation.\n");
+            printf("\tgui                       = Launch the GUI interface.\n");
             printf("\tversion                   = Display %s version.\n", VKCONFIG_NAME);
-            printf("\tlayers                    = Manage system Vulkan Layers.\n");
+            printf("\tloader                    = Manage system Vulkan Loader configurations.\n");
+            printf("\tlayers                    = List Vulkan layers.\n");
             printf("\treset                     = Reset layers configurations.\n");
             printf("\tdoc                       = Create doc files for layer.\n");
             printf("\n");
             printf("  (Use 'vkconfig help <command>' for detailed usage of %s commands.)\n", VKCONFIG_NAME);
-            printf("  (Use 'vkconfig-gui' to open %s graphics interface.)\n", VKCONFIG_NAME);
             break;
         }
         case HELP_HELP: {
@@ -392,23 +503,67 @@ void CommandLine::usage() const {
             printf("\t'layers' - Command to manage system Vulkan Layers\n");
             printf("\n");
             printf("Synopsis\n");
-            printf("\tvkconfig layers (--override | -o) <layers_configuration_file>\n");
-            printf("\tvkconfig layers (--surrender | -s)\n");
             printf("\tvkconfig layers (--list | -l)\n");
             printf("\tvkconfig layers (--list-verbose | -lv)\n");
+            printf("\tvkconfig layers (--path | -p)\n");
+            printf("\tvkconfig layers (--override | -o) <loader_configuration_file>\n");
+            printf("\tvkconfig layers (--surrender | -s)\n");
             printf("\n");
             printf("Description\n");
-            printf("\tvkconfig layers (--override | -o) <layers_configuration_file>\n");
-            printf("\t\tOverride the Vulkan layers using <layers_configuration_file> generated by %s.\n", VKCONFIG_NAME);
-            printf("\n");
-            printf("\tvkconfig layers (--surrender | -s)\n");
-            printf("\t\tSurrender the Vulkan layers control to Vulkan applications.\n");
-            printf("\n");
             printf("\tvkconfig layers (--list | -l)\n");
             printf("\t\tList the Vulkan layers found by %s on the system.\n", VKCONFIG_NAME);
             printf("\n");
             printf("\tvkconfig layers (--list-version | -lv)\n");
             printf("\t\tList the Vulkan layers found by %s on the system with locations and versions.\n", VKCONFIG_NAME);
+            printf("\n");
+            printf("\tvkconfig layers (--path | -p)\n");
+            printf("\t\tList the Vulkan layers locations found by %s on the system.\n", VKCONFIG_NAME);
+            printf("\n");
+            printf("\tvkconfig layers (--override | -o) <loader_configuration_file>\n");
+            printf("\t\tOverride the Vulkan layers using <loader_configuration_file> generated by %s.\n", VKCONFIG_NAME);
+            printf("\t\tDEPRECATED: use `vkconfig loader --override` instead.\n");
+            printf("\n");
+            printf("\tvkconfig layers (--surrender | -s)\n");
+            printf("\t\tSurrender the Vulkan layers control to Vulkan applications.\n");
+            printf("\t\tDEPRECATED: use `vkconfig loader --surrender` instead.\n");
+            break;
+        }
+        case HELP_LOADER: {
+            printf("Name\n");
+            printf("\t'loader' - Command to manage system Vulkan Loader configuration\n");
+            printf("\n");
+            printf("Synopsis\n");
+            printf("\tvkconfig loader (--override | -o) (<configuration_index> | <configuration_name> | <configuration_file>)\n");
+            printf("\tvkconfig loader (--surrender | -s)\n");
+            printf("\tvkconfig loader (--list | -l)\n");
+            printf("\tvkconfig loader (--import | -i) <configuration_file>\n");
+            printf("\tvkconfig loader (--export | -e) (<configuration_index> | <configuration_name>) <configuration_file>\n");
+            printf("\tvkconfig loader (--delete | -d) (<configuration_index> | <configuration_name>)\n");
+            printf("\n");
+            printf("Description\n");
+            printf("\tvkconfig loader (--override | -o) (<configuration_index> | <configuration_name> | <configuration_file>)\n");
+            printf("\t\tOverride the system Vulkan Loader configuration generated by %s.\n", VKCONFIG_NAME);
+            printf("\t\t - <configuration_index> is an index enumerated with `vkconfig loader --list`.\n");
+            printf("\t\t - <configuration_name> is the name of the stored configuration listed with `vkconfig loader --list`.\n");
+            printf("\n");
+            printf("\tvkconfig loader (--surrender | -s)\n");
+            printf("\t\tSurrender the Vulkan Loader configuration to Vulkan applications.\n");
+            printf("\n");
+            printf("\tvkconfig loader (--list | -l)\n");
+            printf("\t\tList the Vulkan Loader configurations found by %s on the system.\n", VKCONFIG_NAME);
+            printf("\n");
+            printf("\tvkconfig loader (--import | -i) <configuration_file>\n");
+            printf("\t\tImport a Vulkan Loader configuration stored by %s on the system.\n", VKCONFIG_NAME);
+            printf("\n");
+            printf("\tvkconfig loader (--export | -e) (<configuration_index> | <configuration_name>) <configuration_file>\n");
+            printf("\t\tExport a Vulkan Loader configuration stored by %s on the system.\n", VKCONFIG_NAME);
+            printf("\t\t - <configuration_index> is an index enumerated with `vkconfig loader --list`.\n");
+            printf("\t\t - <configuration_name> is the name of the stored configuration listed with `vkconfig loader --list`.\n");
+            printf("\n");
+            printf("\tvkconfig loader (--delete | -d) (<configuration_index> | <configuration_name>)\n");
+            printf("\t\tRemove a Vulkan Loader configuration stored by %s on the system.\n", VKCONFIG_NAME);
+            printf("\t\t - <configuration_index> is an index enumerated with `vkconfig loader --list`.\n");
+            printf("\t\t - <configuration_name> is the name of the stored configuration listed with `vkconfig loader --list`.\n");
             break;
         }
         case HELP_DOC: {