+++ vkconfig_gui/tab_applications.cpp
@@ -26,6 +26,7 @@
 #include "../vkconfig_core/alert.h"
 
 #include <QFileDialog>
+#include <QDesktopServices>
 
 TabApplications::TabApplications(MainWindow &window, std::shared_ptr<Ui::MainWindow> ui)
     : Tab(TAB_APPLICATIONS, window, ui), _launch_application(nullptr) {
@@ -36,6 +37,7 @@ TabApplications::TabApplications(MainWindow &window, std::shared_ptr<Ui::MainWin
     this->ui->launch_options_remove->setIcon(::Get(::ICON_OPTIONS_REMOVE));
     this->ui->launch_options_dir_button->setIcon(::Get(::ICON_FOLDER_SEARCH));
     this->ui->launch_options_log_button->setIcon(::Get(::ICON_FILE_SEARCH));
+    this->ui->launch_options_log_open->setIcon(::Get(::ICON_FILE_EXPORT));
 
     this->connect(this->ui->launch_executable_list, SIGNAL(currentIndexChanged(int)), this,
                   SLOT(on_launch_executable_list_activated(int)));
@@ -62,6 +64,7 @@ TabApplications::TabApplications(MainWindow &window, std::shared_ptr<Ui::MainWin
     this->connect(this->ui->launch_options_log_edit, SIGNAL(textEdited(QString)), this,
                   SLOT(on_launch_options_log_textEdited(QString)));
     this->connect(this->ui->launch_options_log_button, SIGNAL(clicked()), this, SLOT(on_launch_options_log_pressed()));
+    this->connect(this->ui->launch_options_log_open, SIGNAL(clicked()), this, SLOT(on_launch_options_log_open_pressed()));
 
     this->connect(this->ui->launch_clear_at_launch, SIGNAL(toggled(bool)), this, SLOT(on_launch_clear_at_launch_toggled(bool)));
     this->connect(this->ui->launch_clear_log, SIGNAL(clicked()), this, SLOT(on_launch_clear_log_pressed()));
@@ -74,9 +77,8 @@ TabApplications::TabApplications(MainWindow &window, std::shared_ptr<Ui::MainWin
     this->ui->launch_log_text->document()->setMaximumBlockCount(65536);
     this->ui->launch_log_text->moveCursor(QTextCursor::End);
 
-    this->ui->launch_options_args_edit->setToolTip("Eg: '--argA --argB'");
-    this->ui->launch_options_envs_edit->setToolTip(VKC_ENV == VKC_ENV_WIN32 ? "Eg: 'ENV_A=ValueA;ENV_B=ValueB;ENV_C='"
-                                                                            : "Eg: 'ENV_A=ValueA:ENV_B=ValueB:ENV_C='");
+    this->ui->launch_options_args_edit->setToolTip("Eg: '--argA --argB=valueB \"--argC=value C\" --argD=\"value D\"'");
+    this->ui->launch_options_envs_edit->setToolTip("Eg: 'ENV_A= ENV_B=ValueB \"ENV_C=Value C\" ENV_D=\"Value D\"'");
 }
 
 TabApplications::~TabApplications() { this->ResetLaunchApplication(); }
@@ -89,6 +91,9 @@ void TabApplications::UpdateUI(UpdateUIMode mode) {
     this->ui->launch_executable_search->setEnabled(!configurator.executables.Empty());
     this->ui->launch_executable_remove->setEnabled(!configurator.executables.Empty());
     this->ui->launch_button->setEnabled(!configurator.executables.Empty());
+
+    const Executable *executable = configurator.executables.GetActiveExecutable();
+    const ExecutableOptions *options = executable->GetActiveOptions();
 
     if (mode == UPDATE_REBUILD_UI) {
         // Rebuild list of applications
@@ -229,7 +234,7 @@ void TabApplications::on_launch_options_list_activated(int index) {
     ui->launch_options_dir_edit->setText(options->working_folder.RelativePath().c_str());
     ui->launch_options_dir_edit->setToolTip(options->working_folder.AbsolutePath().c_str());
     ui->launch_options_args_edit->setText(Merge(options->args, " ").c_str());
-    ui->launch_options_envs_edit->setText(Merge(options->envs, GetToken(PARSE_ENV_VAR)).c_str());
+    ui->launch_options_envs_edit->setText(Merge(options->envs, " ").c_str());
     ui->launch_options_log_edit->setText(options->log_file.RelativePath().c_str());
     ui->launch_options_log_edit->setToolTip(options->log_file.AbsolutePath().c_str());
 }
@@ -320,7 +325,7 @@ void TabApplications::on_launch_options_args_textEdited(const QString &text) {
     Executable *executable = configurator.executables.GetActiveExecutable();
     ExecutableOptions *options = executable->GetActiveOptions();
 
-    options->args = Split(text.toStdString(), " ");
+    options->args = SplitSpace(text.toStdString());
 }
 
 void TabApplications::on_launch_options_envs_textEdited(const QString &text) {
@@ -329,7 +334,7 @@ void TabApplications::on_launch_options_envs_textEdited(const QString &text) {
     Executable *executable = configurator.executables.GetActiveExecutable();
     ExecutableOptions *options = executable->GetActiveOptions();
 
-    options->envs = Split(text.toStdString(), ::GetToken(PARSE_ENV_VAR));
+    options->envs = SplitSpace(text.toStdString());
 }
 
 void TabApplications::on_launch_options_log_textEdited(const QString &text) {
@@ -355,6 +360,19 @@ void TabApplications::on_launch_options_log_pressed() {
         options->log_file = selected_path.toStdString();
         this->ui->launch_options_log_edit->setText(options->log_file.RelativePath().c_str());
     }
+}
+
+void TabApplications::on_launch_options_log_open_pressed() {
+    Configurator &configurator = Configurator::Get();
+
+    Executable *executable = configurator.executables.GetActiveExecutable();
+    ExecutableOptions *options = executable->GetActiveOptions();
+
+    if (!options->log_file.Exists()) {
+        options->log_file.Create(true);
+    }
+
+    QDesktopServices::openUrl(QUrl::fromLocalFile(options->log_file.AbsolutePath().c_str()));
 }
 
 void TabApplications::on_launch_clear_at_launch_toggled(bool checked) {